[refact] #251 : Connect OnboardingView to NamingView
