+++ RelaxOn/Views/Onboarding/OnboardingFinishView.swift
@@ -9,26 +9,25 @@ import SwiftUI
 
 struct OnboardingFinishView: View {
     // MARK: - State Properties
-    @State var userRepositoriesState: [MixedSound] = userRepositories
-    @State var onboardingNavigate: Bool = false
-    @Binding var selectedImageNames: (base: String, melody: String, whiteNoise: String)
-    @Binding var opacityAnimationValues: [Double]
-    @Binding var textEntered: String
+//    @State var userRepositoriesState: [MixedSound] = userRepositories
     @Binding var showOnboarding: Bool
+    
+    // MARK: - General Properties
+    var mixedSound: MixedSound
     
     // MARK: - Life Cycles
     var body: some View {
         ZStack {
-            SelectedImageBackgroundView(selectedImageNames: $selectedImageNames,
-                                        opacityAnimationValues: $opacityAnimationValues)
+            CDCoverImageView(selectedImageNames: mixedSound.getImageName())
+                .toBackground()
+            
             VStack {
                 Spacer()
-                SelectedImageView(framerevise: true,
-                                  selectedImageNames: $selectedImageNames,
-                                  opacityAnimationValues: $opacityAnimationValues)
-                    .frame(width: deviceFrame.screenWidth * 0.6, height: deviceFrame.screenWidth * 0.6)
+                CDCoverImageView(selectedImageNames: mixedSound.getImageName())
+                    .frame(width: deviceFrame.screenWidth * 0.6,
+                           height: deviceFrame.screenWidth * 0.6)
                 
-                Text("\(textEntered)")
+                Text("\(mixedSound.name)")
                     .font(.system(size: 28, weight: .medium))
                     .foregroundColor(.white)
                     .padding()
@@ -44,23 +43,20 @@ struct OnboardingFinishView: View {
 extension OnboardingFinishView {
     @ViewBuilder
     func StartButton() -> some View {
-        NavigationLink(isActive: $onboardingNavigate) {
-            CdLibraryView()
-                .navigationBarHidden(true)
-        } label: {}
-        
         Button {
             UserDefaultsManager.shared.notFirstVisit = true
             showOnboarding = false
-            onboardingNavigate = true
         } label: {
             Text("START")
                 .foregroundColor(.white)
-                .font(.system(size: 20, weight: .medium))
-                .frame(width: deviceFrame.exceptPaddingWidth, height: deviceFrame.screenHeight * 0.07)
-                .cornerRadius(10)
-                .background(.black)
+                .font(.system(size: 20, weight: .light))
+                .frame(width: deviceFrame.exceptPaddingWidth,
+                       height: deviceFrame.screenHeight * 0.07)
+                .background {
+                    RoundedRectangle(cornerRadius: 4)
+                        .foregroundColor(.relaxRealBlack)
+                }
+                .padding()
         }
-        .padding()
     }
 }
